Renamed ThemeManager and implemented editing
Renamed ThemeManager and implemented editing functionality. created an alert for when a theme is created or edited and has less than 2 emojis. Adapted ThemeChooser so that tap gesture on themes while editing takes the index of the theme to edit. For some reason chosenThemeIndex remains 0 until the first theme is selected and ThemeEditor is closed. After that chosenThemeIndex is updated correctly with every selection.