Corrected color bug and restart function
Corrected color bug when initializing one new Theme on first application run. Corrected restart game to use same emojis as in previous game, instead of dealing new cards. Moved dealt logic for dealing animation to EmojiMemoryGame ViewModel, so that cards aren't dealt again when returning to ThemeChooser and opening a Theme with a running game. Changing the Theme forces a new game to be started. Updated game Model so that additional points are given for matching with bonus time remaining.